backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
scikit-learn==1.3.2
joblib==1.4.2
numpy==1.26.4
pandas==2.0.3
PyMySQL==1.1.0
python-dotenv==1.0.1
matplotlib==3.7.2
seaborn==0.12.2
requests==2.32.3
